fix : 간단조회 수정 - 페이지

+++ src/main/java/contactSite/company/PageResponse.java
@@ -3,7 +3,7 @@ package contactSite.company;
 import java.util.List;
 
 public record PageResponse(
-        int totalPage,
+        Long totalPage,
         int pageNumber,
         List<CompanyResponse> companies
 ) {
